requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index -o requirements.txt requirements.in
#
aiohttp-jinja2==1.1.0
aiohttp-remotes==0.1.2
aiohttp==3.5.4
aioredis==1.2.0
async-timeout==3.0.1      # via aiohttp, aioredis
asyncpg==0.18.3
attrs==19.1.0             # via aiohttp
bcrypt==3.1.6
cffi==1.12.2              # via bcrypt
chardet==3.0.4            # via aiohttp
hiredis==1.0.0            # via aioredis
idna==2.8                 # via yarl
jinja2==2.10
markupsafe==1.1.1         # via jinja2
multidict==4.5.2          # via aiohttp, yarl
pycparser==2.19           # via cffi
pyyaml==5.1
six==1.12.0               # via bcrypt
uvloop==0.12.1
yarl==1.3.0               # via aiohttp
